requirements_rag.txt
# RAG System Requirements
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai>=1.30.0
numpy>=1.26.0
python-multipart==0.0.6
python-dotenv==1.0.0
chromadb>=0.5.0
